+++ src/Interfaces/Scalar.php
@@ -1,0 +1,7 @@
+<?php
+namespace MarcoConsiglio\Goniometry\Interfaces;
+
+interface Scalar
+{
+    public function value(): int|float;
+}