+++ CHANGELOG.md
@@ -4,7 +4,9 @@
 - `AngularDistance` class to represent angular distance between two objects.
 - `AngularDistance::createFromDecimal()` method to build an `AngularDistance` object from a `float` or `SexadecimalAngularDistance` type input.
 - `AngularDistance::createFromValues()` method to build an `AngularDistance` object from degrees (`int`), minutes (`int`), seconds (`float`) and `Direction`.
+- `AngularDistance::createFromRadian()` method to build an `AngularDistance` object from a radian value.
 - `SexadecimalAngularDistance` class to represent sexadecimal value of an `AngularDistance` object.
+- `AngularDistanceRadian` class to represent a radian value of an `AngularDistance` object.
 - `AngularDistance::toSexadecimalAngularDistance()` to cast an `AngularDistance` object to a `SexadecimalAngularDistance` object.
 - `AngularDistance::toFloat()` method to cast an `AngularDistance` object to sexadecimal `float` value.
 - `WithAngleFaker::{`